src/requirements.txt
python-dotenv==0.13.0
tweepy==3.8.0
google-cloud-datastore==1.12.0